Enum rusoto_cloudtrail::CreateTrailError

pub enum CreateTrailError {
    CloudWatchLogsDeliveryUnavailable(String),
    InsufficientEncryptionPolicy(String),
    InsufficientS3BucketPolicy(String),
    InsufficientSnsTopicPolicy(String),
    InvalidCloudWatchLogsLogGroupArn(String),
    InvalidCloudWatchLogsRoleArn(String),
    InvalidKmsKeyId(String),
    InvalidParameterCombination(String),
    InvalidS3BucketName(String),
    InvalidS3Prefix(String),
    InvalidSnsTopicName(String),
    InvalidTrailName(String),
    Kms(String),
    KmsKeyDisabled(String),
    KmsKeyNotFound(String),
    MaximumNumberOfTrailsExceeded(String),
    OperationNotPermitted(String),
    S3BucketDoesNotExist(String),
    TrailAlreadyExists(String),
    TrailNotProvided(String),
    UnsupportedOperation(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateTrail

Variants
CloudWatchLogsDeliveryUnavailable(String)

Cannot set a CloudWatch Logs delivery for this region.

InsufficientEncryptionPolicy(String)

This exception is thrown when the policy on the S3 bucket or KMS key is not sufficient.

InsufficientS3BucketPolicy(String)

This exception is thrown when the policy on the S3 bucket is not sufficient.

InsufficientSnsTopicPolicy(String)

This exception is thrown when the policy on the SNS topic is not sufficient.

InvalidCloudWatchLogsLogGroupArn(String)

This exception is thrown when the provided CloudWatch log group is not valid.

InvalidCloudWatchLogsRoleArn(String)

This exception is thrown when the provided role is not valid.

InvalidKmsKeyId(String)

This exception is thrown when the KMS key ARN is invalid.

InvalidParameterCombination(String)

This exception is thrown when the combination of parameters provided is not valid.

InvalidS3BucketName(String)

This exception is thrown when the provided S3 bucket name is not valid.

InvalidS3Prefix(String)

This exception is thrown when the provided S3 prefix is not valid.

InvalidSnsTopicName(String)

This exception is thrown when the provided SNS topic name is not valid.

InvalidTrailName(String)

This exception is thrown when the provided trail name is not valid. Trail names must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid.

Not be in IP address format (for example, 192.168.5.4)

Kms(String)

This exception is thrown when there is an issue with the specified KMS key and the trail can’t be updated.

KmsKeyDisabled(String)

This exception is deprecated.

KmsKeyNotFound(String)

This exception is thrown when the KMS key does not exist, or when the S3 bucket and the KMS key are not in the same region.

MaximumNumberOfTrailsExceeded(String)

This exception is thrown when the maximum number of trails is reached.

OperationNotPermitted(String)

This exception is thrown when the requested operation is not permitted.

S3BucketDoesNotExist(String)

This exception is thrown when the specified S3 bucket does not exist.

TrailAlreadyExists(String)

This exception is thrown when the specified trail already exists.

TrailNotProvided(String)

This exception is deprecated.

UnsupportedOperation(String)

This exception is thrown when the requested operation is not supported.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateTrailError

fn from_body(body: &str) -> CreateTrailError