Enum rusoto_cloudtrail::UpdateTrailError

pub enum UpdateTrailError {
    CloudTrailAccessNotEnabled(String),
    CloudWatchLogsDeliveryUnavailable(String),
    InsufficientDependencyServiceAccessPermission(String),
    InsufficientEncryptionPolicy(String),
    InsufficientS3BucketPolicy(String),
    InsufficientSnsTopicPolicy(String),
    InvalidCloudWatchLogsLogGroupArn(String),
    InvalidCloudWatchLogsRoleArn(String),
    InvalidHomeRegion(String),
    InvalidKmsKeyId(String),
    InvalidParameterCombination(String),
    InvalidS3BucketName(String),
    InvalidS3Prefix(String),
    InvalidSnsTopicName(String),
    InvalidTrailName(String),
    Kms(String),
    KmsKeyDisabled(String),
    KmsKeyNotFound(String),
    NotOrganizationMasterAccount(String),
    OperationNotPermitted(String),
    OrganizationNotInAllFeaturesMode(String),
    OrganizationsNotInUse(String),
    S3BucketDoesNotExist(String),
    TrailNotFound(String),
    TrailNotProvided(String),
    UnsupportedOperation(String),
}

Errors returned by UpdateTrail

Variants
CloudTrailAccessNotEnabled(String)

This exception is thrown when trusted access has not been enabled between AWS CloudTrail and AWS Organizations. For more information, see Enabling Trusted Access with Other AWS Services and Prepare For Creating a Trail For Your Organization.

CloudWatchLogsDeliveryUnavailable(String)

Cannot set a CloudWatch Logs delivery for this region.

InsufficientDependencyServiceAccessPermission(String)

This exception is thrown when the IAM user or role that is used to create the organization trail is lacking one or more required permissions for creating an organization trail in a required service. For more information, see Prepare For Creating a Trail For Your Organization.

InsufficientEncryptionPolicy(String)

This exception is thrown when the policy on the S3 bucket or KMS key is not sufficient.

InsufficientS3BucketPolicy(String)

This exception is thrown when the policy on the S3 bucket is not sufficient.

InsufficientSnsTopicPolicy(String)

This exception is thrown when the policy on the SNS topic is not sufficient.

InvalidCloudWatchLogsLogGroupArn(String)

This exception is thrown when the provided CloudWatch log group is not valid.

InvalidCloudWatchLogsRoleArn(String)

This exception is thrown when the provided role is not valid.

InvalidHomeRegion(String)

This exception is thrown when an operation is called on a trail from a region other than the region in which the trail was created.

InvalidKmsKeyId(String)

This exception is thrown when the KMS key ARN is invalid.

InvalidParameterCombination(String)

This exception is thrown when the combination of parameters provided is not valid.

InvalidS3BucketName(String)

This exception is thrown when the provided S3 bucket name is not valid.

InvalidS3Prefix(String)

This exception is thrown when the provided S3 prefix is not valid.

InvalidSnsTopicName(String)

This exception is thrown when the provided SNS topic name is not valid.

InvalidTrailName(String)

This exception is thrown when the provided trail name is not valid. Trail names must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-namespace and my--namespace are invalid.

Not be in IP address format (for example, 192.168.5.4)

Kms(String)

This exception is thrown when there is an issue with the specified KMS key and the trail can’t be updated.

KmsKeyDisabled(String)

This exception is deprecated.

KmsKeyNotFound(String)

This exception is thrown when the KMS key does not exist, or when the S3 bucket and the KMS key are not in the same region.

NotOrganizationMasterAccount(String)

This exception is thrown when the AWS account making the request to create or update an organization trail is not the master account for an organization in AWS Organizations. For more information, see Prepare For Creating a Trail For Your Organization.

OperationNotPermitted(String)

This exception is thrown when the requested operation is not permitted.

OrganizationNotInAllFeaturesMode(String)

This exception is thrown when AWS Organizations is not configured to support all features. All features must be enabled in AWS Organization to support creating an organization trail. For more information, see Prepare For Creating a Trail For Your Organization.

OrganizationsNotInUse(String)

This exception is thrown when the request is made from an AWS account that is not a member of an organization. To make this request, sign in using the credentials of an account that belongs to an organization.

S3BucketDoesNotExist(String)

This exception is thrown when the specified S3 bucket does not exist.

TrailNotFound(String)

This exception is thrown when the trail with the given name is not found.

TrailNotProvided(String)

This exception is deprecated.

UnsupportedOperation(String)

This exception is thrown when the requested operation is not supported.

Methods

impl UpdateTrailError

pub fn from_response(res: BufferedHttpResponse) -> RusotoError<UpdateTrailError>