Struct rusoto_cloudtrail::GetTrailStatusResponse

pub struct GetTrailStatusResponse {
    pub is_logging: Option<bool>,
    pub latest_cloud_watch_logs_delivery_error: Option<String>,
    pub latest_cloud_watch_logs_delivery_time: Option<f64>,
    pub latest_delivery_attempt_succeeded: Option<String>,
    pub latest_delivery_attempt_time: Option<String>,
    pub latest_delivery_error: Option<String>,
    pub latest_delivery_time: Option<f64>,
    pub latest_digest_delivery_error: Option<String>,
    pub latest_digest_delivery_time: Option<f64>,
    pub latest_notification_attempt_succeeded: Option<String>,
    pub latest_notification_attempt_time: Option<String>,
    pub latest_notification_error: Option<String>,
    pub latest_notification_time: Option<f64>,
    pub start_logging_time: Option<f64>,
    pub stop_logging_time: Option<f64>,
    pub time_logging_started: Option<String>,
    pub time_logging_stopped: Option<String>,
}

Returns the objects or data listed below if successful. Otherwise, returns an error.

Fields
is_logging: Option<bool>

Whether the CloudTrail is currently logging AWS API calls.

latest_cloud_watch_logs_delivery_error: Option<String>

Displays any CloudWatch Logs error that CloudTrail encountered when attempting to deliver logs to CloudWatch Logs.

latest_cloud_watch_logs_delivery_time: Option<f64>

Displays the most recent date and time when CloudTrail delivered logs to CloudWatch Logs.

latest_delivery_attempt_succeeded: Option<String>

This field is deprecated.

latest_delivery_attempt_time: Option<String>

This field is deprecated.

latest_delivery_error: Option<String>

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver log files to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket and will not occur for timeouts. To resolve the issue, create a new bucket and call UpdateTrail to specify the new bucket, or fix the existing objects so that CloudTrail can again write to the bucket.

latest_delivery_time: Option<f64>

Specifies the date and time that CloudTrail last delivered log files to an account's Amazon S3 bucket.

latest_digest_delivery_error: Option<String>

Displays any Amazon S3 error that CloudTrail encountered when attempting to deliver a digest file to the designated bucket. For more information see the topic Error Responses in the Amazon S3 API Reference.

This error occurs only when there is a problem with the destination S3 bucket and will not occur for timeouts. To resolve the issue, create a new bucket and call UpdateTrail to specify the new bucket, or fix the existing objects so that CloudTrail can again write to the bucket.

latest_digest_delivery_time: Option<f64>

Specifies the date and time that CloudTrail last delivered a digest file to an account's Amazon S3 bucket.

latest_notification_attempt_succeeded: Option<String>

This field is deprecated.

latest_notification_attempt_time: Option<String>

This field is deprecated.

latest_notification_error: Option<String>

Displays any Amazon SNS error that CloudTrail encountered when attempting to send a notification. For more information about Amazon SNS errors, see the Amazon SNS Developer Guide.

latest_notification_time: Option<f64>

Specifies the date and time of the most recent Amazon SNS notification that CloudTrail has written a new log file to an account's Amazon S3 bucket.

start_logging_time: Option<f64>

Specifies the most recent date and time when CloudTrail started recording API calls for an AWS account.

stop_logging_time: Option<f64>

Specifies the most recent date and time when CloudTrail stopped recording API calls for an AWS account.

time_logging_started: Option<String>

This field is deprecated.

time_logging_stopped: Option<String>

This field is deprecated.