Crate rusoto_cloudtrail

AWS CloudTrail
This is the CloudTrail API Reference. It provides descriptions of actions, data types, common parameters, and common errors for CloudTrail.

CloudTrail is a web service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. The recorded information includes the identity of the user, the start time of the AWS API call, the source IP address, the request parameters, and the response elements returned by the service.

As an alternative to the API, you can use one of the AWS SDKs, which consist of libraries and sample code for various programming languages and platforms (Java, Ruby, .NET, iOS, Android, etc.). The SDKs provide a convenient way to create programmatic access to AWSCloudTrail. For example, the SDKs take care of cryptographically signing requests, managing errors, and retrying requests automatically. For information about the AWS SDKs, including how to download and install them, see the Tools for Amazon Web Services page.

See the AWS CloudTrail User Guide for information about the data that is included with each AWS API call listed in the log files.

If you're using the service, you're probably looking for CloudTrailClient and CloudTrail.

Structs

AddTagsRequest
Specifies the tags to add to a trail.

AddTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

CloudTrailClient
A client for the CloudTrail API.

CreateTrailRequest
Specifies the settings for each trail.

CreateTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

DataResource

The Amazon S3 objects that you specify in your event selectors for your trail to log data events. Data events are object-level API operations that access S3 objects, such as GetObject, DeleteObject, and PutObject. You can specify up to 250 S3 buckets and object prefixes for a trail.

Example

You create an event selector for a trail and specify an S3 bucket and an empty prefix, such as arn:aws:s3:::bucket-1/.

You upload an image file to bucket-1.

The PutObject API operation occurs on an object in the S3 bucket that you specified in the event selector. The trail processes and logs the event.

You upload another image file to a different S3 bucket named arn:aws:s3:::bucket-2.

The event occurs on an object in an S3 bucket that you didn't specify in the event selector. The trail doesn’t log the event.

DeleteTrailRequest
The request that specifies the name of a trail to delete.

DeleteTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

DescribeTrailsRequest
Returns information about the trail.

DescribeTrailsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Event
Contains information about an event that was returned by a lookup request. The result includes a representation of a CloudTrail event.

EventSelector
Use event selectors to specify whether you want your trail to log management and/or data events. When an event occurs in your account, CloudTrail evaluates the event selector for all trails. For each trail, if the event matches any event selector, the trail processes and logs the event. If the event doesn't match any event selector, the trail doesn't log the event.

You can configure up to five event selectors for a trail.

GetEventSelectorsRequest

GetEventSelectorsResponse

GetTrailStatusRequest
The name of a trail about which you want the current status.

GetTrailStatusResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

ListPublicKeysRequest
Requests the public keys for a specified time range.

ListPublicKeysResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

ListTagsRequest
Specifies a list of trail tags to return.

ListTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

LookupAttribute
Specifies an attribute and value that filter the events returned.

LookupEventsRequest
Contains a request for LookupEvents.

LookupEventsResponse
Contains a response to a LookupEvents action.

PublicKey
Contains information about a returned public key.

PutEventSelectorsRequest

PutEventSelectorsResponse

RemoveTagsRequest
Specifies the tags to remove from a trail.

RemoveTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Resource
Specifies the type and name of a resource referenced by an event.

ResourceTag
A resource tag.

StartLoggingRequest
The request to CloudTrail to start logging AWS API calls for an account.

StartLoggingResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

StopLoggingRequest
Passes the request to CloudTrail to stop logging AWS API calls for the specified account.

StopLoggingResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Tag
A custom key-value pair associated with a resource such as a CloudTrail trail.

Trail
The settings for a trail.

UpdateTrailRequest
Specifies settings to update for the trail.

UpdateTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Enums

AddTagsError
Errors returned by AddTags

CreateTrailError
Errors returned by CreateTrail

DeleteTrailError
Errors returned by DeleteTrail

DescribeTrailsError
Errors returned by DescribeTrails

GetEventSelectorsError
Errors returned by GetEventSelectors

GetTrailStatusError
Errors returned by GetTrailStatus

ListPublicKeysError
Errors returned by ListPublicKeys

ListTagsError
Errors returned by ListTags

LookupEventsError
Errors returned by LookupEvents

PutEventSelectorsError
Errors returned by PutEventSelectors

RemoveTagsError
Errors returned by RemoveTags

StartLoggingError
Errors returned by StartLogging

StopLoggingError
Errors returned by StopLogging

UpdateTrailError
Errors returned by UpdateTrail

Traits

CloudTrail
Trait representing the capabilities of the CloudTrail API. CloudTrail clients implement this trait.