Struct rusoto_cloudtrail::CloudTrailClient

pub struct CloudTrailClient { /* fields omitted */ }

A client for the CloudTrail API.

Methods

impl CloudTrailClient

pub fn new(region: Region) -> CloudTrailClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudTrailClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,